+++ test/functional/test_framework/wallet.py
@@ -13,8 +13,6 @@ from typing import (
     Optional,
 )
 from test_framework.address import (
-    base58_to_byte,
-    bech32_to_bytes,
     create_deterministic_address_bcrt1_p2tr_op_true,
     key_to_p2pkh,
     key_to_p2sh_p2wpkh,
@@ -49,9 +47,6 @@ from test_framework.script_util import (
     key_to_p2pkh_script,
     key_to_p2sh_p2wpkh_script,
     key_to_p2wpkh_script,
-    keyhash_to_p2pkh_script,
-    program_to_witness_script,
-    scripthash_to_p2sh_script,
 )
 from test_framework.util import (
     assert_equal,
@@ -412,18 +407,3 @@ def getnewdestination(address_type='bech32m'):
     else:
         assert False
     return pubkey, scriptpubkey, address
-
-
-def address_to_scriptpubkey(address):
-    """Converts a given address to the corresponding output script (scriptPubKey)."""
-    version, payload = bech32_to_bytes(address)
-    if version is not None:
-        return program_to_witness_script(version, payload) # testnet segwit scriptpubkey
-    payload, version = base58_to_byte(address)
-    if version == 111:  # testnet pubkey hash
-        return keyhash_to_p2pkh_script(payload)
-    elif version == 196:  # testnet script hash
-        return scripthash_to_p2sh_script(payload)
-    # TODO: also support other address formats
-    else:
-        assert False
